Refactor: Move file operations to dedicated file
